token now also has ints instead of longs just like the lexer uses.

+++ Lexing/Token.cs
@@ -4,17 +4,17 @@
     {
 
         public TokenType TokenType { get; set; }
-        public long Line { get; init; }
-        public long Column { get; init; }
+        public int Line { get; init; }
+        public int Column { get; init; }
         public string Name { get; init; }
         public string? StringValue { get; set; }
         public float? FloatValue { get; set; } // whats better, float or double?
         public long? IntegerValue { get; set; } // todo: how to handle longs?
         public bool? BooleanValue { get; set; }
-        public Token(TokenType tokenType, long line, long column) : this(tokenType, tokenType.ToString(), line, column) { }
-        public Token(long line, long column) : this(default, ((TokenType)default).ToString(), line, column) { }
-        public Token(TokenType tokenType, char name, long line, long column) : this(tokenType, name.ToString(), line, column) { }
-        public Token(TokenType tokenType, string name, long line, long column)
+        public Token(TokenType tokenType, int line, int column) : this(tokenType, tokenType.ToString(), line, column) { }
+        public Token(int line, int column) : this(default, ((TokenType)default).ToString(), line, column) { }
+        public Token(TokenType tokenType, char name, int line, int column) : this(tokenType, name.ToString(), line, column) { }
+        public Token(TokenType tokenType, string name, int line, int column)
         {
             TokenType = tokenType;
             Line = line;
